+++ src/state/features/application/actions.ts
@@ -75,6 +75,22 @@ const getCompanyDetailsWithAddress = createAsyncThunk(
   }
 )
 
+const getUniqueIdentifier = createAsyncThunk(
+  `${name}/uniqueIdentifier/get`,
+  async (alpha2Code: string) => {
+    try {
+      return await ApplicationApi.getInstance().getUniqueIdentifier(
+        alpha2Code
+      )
+    } catch (error: unknown) {
+      console.error('api call error:', error)
+      throw Error(
+        'No identifier found.'
+      )
+    }
+  }
+)
+
 const saveCompanyDetailsWithAddress = createAsyncThunk(
   `${name}/companyDetailsWithAddress/save-address`,
   async ({
@@ -103,5 +119,6 @@ export {
   updateInvitation,
   updateStatus,
   getCompanyDetailsWithAddress,
+  getUniqueIdentifier,
   saveCompanyDetailsWithAddress,
 }